+++ lib/domain/usecases/login_with_google_usecase.dart
@@ -1,0 +1,14 @@
+import 'package:dartz/dartz.dart';
+import '../../core/error/failures.dart';
+import '../entities/user.dart';
+import '../repositories/auth_repository.dart';
+
+class LoginWithGoogleUseCase {
+  final AuthRepository repository;
+
+  LoginWithGoogleUseCase(this.repository);
+
+  Future<Either<Failure, User>> call() async {
+    return await repository.signInWithGoogle();
+  }
+}